requirements.txt
# Core dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0
openai>=1.12.0
groundx-python-sdk>=0.1.0
redis>=5.0.0

# Common dependencies that might be needed
numpy>=1.26.0
pandas>=2.2.0
python-dateutil>=2.8.2
requests>=2.31.0
typing-extensions>=4.9.0
urllib3>=2.2.0
aiohttp>=3.9.0
hashlib>=20081119

# Optional for production
python-json-logger>=2.0.7  # For better JSON formatted logging
tenacity>=8.2.0  # For retrying failed API calls